Struct google_adexchangebuyer2_v2_beta1::api::AccountMethods

pub struct AccountMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on account resources. It is not used directly, but through the AdExchangeBuyerII hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangebuyer2_v2_beta1 as adexchangebuyer2_v2_beta1;
 
use std::default::Default;
use oauth2;
use adexchangebuyer2_v2_beta1::AdExchangeBuyerII;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = AdExchangeBuyerII::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `clients_create(...)`, `clients_get(...)`, `clients_invitations_create(...)`, `clients_invitations_get(...)`, `clients_invitations_list(...)`, `clients_list(...)`, `clients_update(...)`, `clients_users_get(...)`, `clients_users_list(...)`, `clients_users_update(...)`, `creatives_create(...)`, `creatives_deal_associations_add(...)`, `creatives_deal_associations_list(...)`, `creatives_deal_associations_remove(...)`, `creatives_get(...)`, `creatives_list(...)`, `creatives_stop_watching(...)`, `creatives_update(...)`, `creatives_watch(...)`, `finalized_proposals_list(...)`, `products_get(...)`, `products_list(...)`, `proposals_accept(...)`, `proposals_add_note(...)`, `proposals_cancel_negotiation(...)`, `proposals_complete_setup(...)`, `proposals_create(...)`, `proposals_get(...)`, `proposals_list(...)`, `proposals_pause(...)`, `proposals_resume(...)`, `proposals_update(...)`, `publisher_profiles_get(...)` and `publisher_profiles_list(...)`
// to build up your call.
let rb = hub.accounts();

Implementations

impl<'a, C> AccountMethods<'a, C>

pub fn clients_invitations_create(
&self,
request: ClientUserInvitation,
account_id: &str,
client_account_id: &str
) -> AccountClientInvitationCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates and sends out an email invitation to access an Ad Exchange client buyer account.

Arguments

request - No description provided.

accountId - Numerical account ID of the client's sponsor buyer. (required)

clientAccountId - Numerical account ID of the client buyer that the user should be associated with. (required)

pub fn clients_invitations_get(
&self,
account_id: &str,
client_account_id: &str,
invitation_id: &str
) -> AccountClientInvitationGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves an existing client user invitation.

Arguments

accountId - Numerical account ID of the client's sponsor buyer. (required)

clientAccountId - Numerical account ID of the client buyer that the user invitation to be retrieved is associated with. (required)

invitationId - Numerical identifier of the user invitation to retrieve. (required)

pub fn clients_invitations_list(
&self,
account_id: &str,
client_account_id: &str
) -> AccountClientInvitationListCall<'a, C>

Create a builder to help you perform the following task:

Lists all the client users invitations for a client with a given account ID.

Arguments

accountId - Numerical account ID of the client's sponsor buyer. (required)

clientAccountId - Numerical account ID of the client buyer to list invitations for. (required) You must either specify a string representation of a numerical account identifier or the - character to list all the invitations for all the clients of a given sponsor buyer.

pub fn clients_users_get(
&self,
account_id: &str,
client_account_id: &str,
user_id: &str
) -> AccountClientUserGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves an existing client user.

Arguments

accountId - Numerical account ID of the client's sponsor buyer. (required)

clientAccountId - Numerical account ID of the client buyer that the user to be retrieved is associated with. (required)

userId - Numerical identifier of the user to retrieve. (required)

pub fn clients_users_list(
&self,
account_id: &str,
client_account_id: &str
) -> AccountClientUserListCall<'a, C>

Create a builder to help you perform the following task:

Lists all the known client users for a specified sponsor buyer account ID.

Arguments

accountId - Numerical account ID of the sponsor buyer of the client to list users for. (required)

clientAccountId - The account ID of the client buyer to list users for. (required) You must specify either a string representation of a numerical account identifier or the - character to list all the client users for all the clients of a given sponsor buyer.

pub fn clients_users_update(
&self,
request: ClientUser,
account_id: &str,
client_account_id: &str,
user_id: &str
) -> AccountClientUserUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing client user. Only the user status can be changed on update.

Arguments

request - No description provided.

accountId - Numerical account ID of the client's sponsor buyer. (required)

clientAccountId - Numerical account ID of the client buyer that the user to be retrieved is associated with. (required)

userId - Numerical identifier of the user to retrieve. (required)

pub fn clients_create(
&self,
request: Client,
account_id: &str
) -> AccountClientCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new client buyer.

Arguments

request - No description provided.

accountId - Unique numerical account ID for the buyer of which the client buyer is a customer; the sponsor buyer to create a client for. (required)

pub fn clients_get(
&self,
account_id: &str,
client_account_id: &str
) -> AccountClientGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a client buyer with a given client account ID.

Arguments

accountId - Numerical account ID of the client's sponsor buyer. (required)

clientAccountId - Numerical account ID of the client buyer to retrieve. (required)

pub fn clients_list(&self, account_id: &str) -> AccountClientListCall<'a, C>

Create a builder to help you perform the following task:

Lists all the clients for the current sponsor buyer.

Arguments

accountId - Unique numerical account ID of the sponsor buyer to list the clients for.

pub fn clients_update(
&self,
request: Client,
account_id: &str,
client_account_id: &str
) -> AccountClientUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing client buyer.

Arguments

request - No description provided.

accountId - Unique numerical account ID for the buyer of which the client buyer is a customer; the sponsor buyer to update a client for. (required)

clientAccountId - Unique numerical account ID of the client to update. (required)

pub fn creatives_deal_associations_add(
&self,
request: AddDealAssociationRequest,
account_id: &str,
creative_id: &str
) -> AccountCreativeDealAssociationAddCall<'a, C>

Create a builder to help you perform the following task:

Associate an existing deal with a creative.

Arguments

request - No description provided.

accountId - The account the creative belongs to.

creativeId - The ID of the creative associated with the deal.

pub fn creatives_deal_associations_list(
&self,
account_id: &str,
creative_id: &str
) -> AccountCreativeDealAssociationListCall<'a, C>

Create a builder to help you perform the following task:

List all creative-deal associations.

Arguments

accountId - The account to list the associations from. Specify "-" to list all creatives the current user has access to.

creativeId - The creative ID to list the associations from. Specify "-" to list all creatives under the above account.

pub fn creatives_deal_associations_remove(
&self,
request: RemoveDealAssociationRequest,
account_id: &str,
creative_id: &str
) -> AccountCreativeDealAssociationRemoveCall<'a, C>

Create a builder to help you perform the following task:

Remove the association between a deal and a creative.

Arguments

request - No description provided.

accountId - The account the creative belongs to.

creativeId - The ID of the creative associated with the deal.

pub fn creatives_create(
&self,
request: Creative,
account_id: &str
) -> AccountCreativeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a creative.

Arguments

request - No description provided.

accountId - The account that this creative belongs to. Can be used to filter the response of the creatives.list method.

pub fn creatives_get(
&self,
account_id: &str,
creative_id: &str
) -> AccountCreativeGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a creative.

Arguments

accountId - The account the creative belongs to.

creativeId - The ID of the creative to retrieve.

pub fn creatives_list(&self, account_id: &str) -> AccountCreativeListCall<'a, C>

Create a builder to help you perform the following task:

Lists creatives.

Arguments

accountId - The account to list the creatives from. Specify "-" to list all creatives the current user has access to.

pub fn creatives_stop_watching(
&self,
request: StopWatchingCreativeRequest,
account_id: &str,
creative_id: &str
) -> AccountCreativeStopWatchingCall<'a, C>

Create a builder to help you perform the following task:

Stops watching a creative. Will stop push notifications being sent to the topics when the creative changes status.

Arguments

request - No description provided.

accountId - The account of the creative to stop notifications for.

creativeId - The creative ID of the creative to stop notifications for. Specify "-" to specify stopping account level notifications.

pub fn creatives_update(
&self,
request: Creative,
account_id: &str,
creative_id: &str
) -> AccountCreativeUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates a creative.

Arguments

request - No description provided.

accountId - The account that this creative belongs to. Can be used to filter the response of the creatives.list method.

creativeId - The buyer-defined creative ID of this creative. Can be used to filter the response of the creatives.list method.

pub fn creatives_watch(
&self,
request: WatchCreativeRequest,
account_id: &str,
creative_id: &str
) -> AccountCreativeWatchCall<'a, C>

Create a builder to help you perform the following task:

Watches a creative. Will result in push notifications being sent to the topic when the creative changes status.

Arguments

request - No description provided.

accountId - The account of the creative to watch.

creativeId - The creative ID to watch for status changes. Specify "-" to watch all creatives under the above account. If both creative-level and account-level notifications are sent, only a single notification will be sent to the creative-level notification topic.

pub fn finalized_proposals_list(
&self,
account_id: &str
) -> AccountFinalizedProposalListCall<'a, C>

Create a builder to help you perform the following task:

List finalized proposals, regardless if a proposal is being renegotiated. A filter expression (PQL query) may be specified to filter the results. The notes will not be returned.

Arguments

accountId - Account ID of the buyer.

pub fn products_get(
&self,
account_id: &str,
product_id: &str
) -> AccountProductGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the requested product by ID.

Arguments

accountId - Account ID of the buyer.

productId - The ID for the product to get the head revision for.

pub fn products_list(&self, account_id: &str) -> AccountProductListCall<'a, C>

Create a builder to help you perform the following task:

List all products visible to the buyer (optionally filtered by the specified PQL query).

Arguments

accountId - Account ID of the buyer.

pub fn proposals_accept(
&self,
request: AcceptProposalRequest,
account_id: &str,
proposal_id: &str
) -> AccountProposalAcceptCall<'a, C>

Create a builder to help you perform the following task:

Mark the proposal as accepted at the given revision number. If the number does not match the server's revision number an ABORTED error message will be returned. This call updates the proposal_state from PROPOSED to BUYER_ACCEPTED, or from SELLER_ACCEPTED to FINALIZED. Upon calling this endpoint, the buyer implicitly agrees to the terms and conditions optionally set within the proposal by the publisher.

Arguments

request - No description provided.

accountId - Account ID of the buyer.

proposalId - The ID of the proposal to accept.

pub fn proposals_add_note(
&self,
request: AddNoteRequest,
account_id: &str,
proposal_id: &str
) -> AccountProposalAddNoteCall<'a, C>

Create a builder to help you perform the following task:

Create a new note and attach it to the proposal. The note is assigned a unique ID by the server. The proposal revision number will not increase when associated with a new note.

Arguments

request - No description provided.

accountId - Account ID of the buyer.

proposalId - The ID of the proposal to attach the note to.

pub fn proposals_cancel_negotiation(
&self,
request: CancelNegotiationRequest,
account_id: &str,
proposal_id: &str
) -> AccountProposalCancelNegotiationCall<'a, C>

Create a builder to help you perform the following task:

Cancel an ongoing negotiation on a proposal. This does not cancel or end serving for the deals if the proposal has been finalized, but only cancels a negotiation unilaterally.

Arguments

request - No description provided.

accountId - Account ID of the buyer.

proposalId - The ID of the proposal to cancel negotiation for.

pub fn proposals_complete_setup(
&self,
request: CompleteSetupRequest,
account_id: &str,
proposal_id: &str
) -> AccountProposalCompleteSetupCall<'a, C>

Create a builder to help you perform the following task:

Update the given proposal to indicate that setup has been completed. This method is called by the buyer when the line items have been created on their end for a finalized proposal and all the required creatives have been uploaded using the creatives API. This call updates the is_setup_completed bit on the proposal and also notifies the seller. The server will advance the revision number of the most recent proposal.

Arguments

request - No description provided.

accountId - Account ID of the buyer.

proposalId - The ID of the proposal to mark as setup completed.

pub fn proposals_create(
&self,
request: Proposal,
account_id: &str
) -> AccountProposalCreateCall<'a, C>

Create a builder to help you perform the following task:

Create the given proposal. Each created proposal and any deals it contains are assigned a unique ID by the server.

Arguments

request - No description provided.

accountId - Account ID of the buyer.

pub fn proposals_get(
&self,
account_id: &str,
proposal_id: &str
) -> AccountProposalGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a proposal given its ID. The proposal is returned at its head revision.

Arguments

accountId - Account ID of the buyer.

proposalId - The unique ID of the proposal

pub fn proposals_list(&self, account_id: &str) -> AccountProposalListCall<'a, C>

Create a builder to help you perform the following task:

List proposals. A filter expression (PQL query) may be specified to filter the results. To retrieve all finalized proposals, regardless if a proposal is being renegotiated, see the FinalizedProposals resource. Note that Bidder/ChildSeat relationships differ from the usual behavior. A Bidder account can only see its child seats' proposals by specifying the ChildSeat's accountId in the request path.

Arguments

accountId - Account ID of the buyer.

pub fn proposals_pause(
&self,
request: PauseProposalRequest,
account_id: &str,
proposal_id: &str
) -> AccountProposalPauseCall<'a, C>

Create a builder to help you perform the following task:

Update the given proposal to pause serving. This method will set the DealServingMetadata.DealPauseStatus.has_buyer_paused bit to true for all deals in the proposal. It is a no-op to pause an already-paused proposal. It is an error to call PauseProposal for a proposal that is not finalized or renegotiating.

Arguments

request - No description provided.

accountId - Account ID of the buyer.

proposalId - The ID of the proposal to pause.

pub fn proposals_resume(
&self,
request: ResumeProposalRequest,
account_id: &str,
proposal_id: &str
) -> AccountProposalResumeCall<'a, C>

Create a builder to help you perform the following task:

Update the given proposal to resume serving. This method will set the DealServingMetadata.DealPauseStatus.has_buyer_paused bit to false for all deals in the proposal. Note that if the has_seller_paused bit is also set, serving will not resume until the seller also resumes. It is a no-op to resume an already-running proposal. It is an error to call ResumeProposal for a proposal that is not finalized or renegotiating.

Arguments

request - No description provided.

accountId - Account ID of the buyer.

proposalId - The ID of the proposal to resume.

pub fn proposals_update(
&self,
request: Proposal,
account_id: &str,
proposal_id: &str
) -> AccountProposalUpdateCall<'a, C>

Create a builder to help you perform the following task:

Update the given proposal at the client known revision number. If the server revision has advanced since the passed-in proposal.proposal_revision, an ABORTED error message will be returned. Only the buyer-modifiable fields of the proposal will be updated. Note that the deals in the proposal will be updated to match the passed-in copy. If a passed-in deal does not have a deal_id, the server will assign a new unique ID and create the deal. If passed-in deal has a deal_id, it will be updated to match the passed-in copy. Any existing deals not present in the passed-in proposal will be deleted. It is an error to pass in a deal with a deal_id not present at head.

Arguments

request - No description provided.

accountId - Account ID of the buyer.

proposalId - The unique ID of the proposal.

pub fn publisher_profiles_get(
&self,
account_id: &str,
publisher_profile_id: &str
) -> AccountPublisherProfileGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the requested publisher profile by id.

Arguments

accountId - Account ID of the buyer.

publisherProfileId - The id for the publisher profile to get.

pub fn publisher_profiles_list(
&self,
account_id: &str
) -> AccountPublisherProfileListCall<'a, C>

Create a builder to help you perform the following task:

List all publisher profiles visible to the buyer

Arguments

accountId - Account ID of the buyer.