Struct google_adexchangebuyer2_v2_beta1::api::Deal

pub struct Deal {
    pub available_end_time: Option<String>,
    pub available_start_time: Option<String>,
    pub buyer_private_data: Option<PrivateData>,
    pub create_product_id: Option<String>,
    pub create_product_revision: Option<String>,
    pub create_time: Option<String>,
    pub creative_pre_approval_policy: Option<String>,
    pub creative_restrictions: Option<CreativeRestrictions>,
    pub creative_safe_frame_compatibility: Option<String>,
    pub deal_id: Option<String>,
    pub deal_serving_metadata: Option<DealServingMetadata>,
    pub deal_terms: Option<DealTerms>,
    pub delivery_control: Option<DeliveryControl>,
    pub description: Option<String>,
    pub display_name: Option<String>,
    pub external_deal_id: Option<String>,
    pub is_setup_complete: Option<bool>,
    pub programmatic_creative_source: Option<String>,
    pub proposal_id: Option<String>,
    pub seller_contacts: Option<Vec<ContactInformation>>,
    pub syndication_product: Option<String>,
    pub targeting: Option<MarketplaceTargeting>,
    pub targeting_criterion: Option<Vec<TargetingCriteria>>,
    pub update_time: Option<String>,
    pub web_property_code: Option<String>,
}

A deal represents a segment of inventory for displaying ads on. A proposal can contain multiple deals. A deal contains the terms and targeting information that is used for serving.

This type is not used in any activity, and only used as part of another schema.

Fields
available_end_time: Option<String>

Proposed flight end time of the deal. This will generally be stored in a granularity of a second. A value is not required for Private Auction deals or Preferred Deals.

available_start_time: Option<String>

Optional. Proposed flight start time of the deal. This will generally be stored in the granularity of one second since deal serving starts at seconds boundary. Any time specified with more granularity (e.g., in milliseconds) will be truncated towards the start of time in seconds.

buyer_private_data: Option<PrivateData>

Buyer private data (hidden from seller).

create_product_id: Option<String>

The product ID from which this deal was created. Note: This field may be set only when creating the resource. Modifying this field while updating the resource will result in an error.

create_product_revision: Option<String>

Optional. Revision number of the product that the deal was created from. If present on create, and the server product_revision has advanced since the passed-in create_product_revision, an ABORTED error will be returned. Note: This field may be set only when creating the resource. Modifying this field while updating the resource will result in an error.

create_time: Option<String>

Output only. The time of the deal creation.

creative_pre_approval_policy: Option<String>

Output only. Specifies the creative pre-approval policy.

creative_restrictions: Option<CreativeRestrictions>

Output only. Restricitions about the creatives associated with the deal (i.e., size) This is available for Programmatic Guaranteed/Preferred Deals in Ad Manager.

creative_safe_frame_compatibility: Option<String>

Output only. Specifies whether the creative is safeFrame compatible.

deal_id: Option<String>

Output only. A unique deal ID for the deal (server-assigned).

deal_serving_metadata: Option<DealServingMetadata>

Output only. Metadata about the serving status of this deal.

deal_terms: Option<DealTerms>

The negotiable terms of the deal.

delivery_control: Option<DeliveryControl>

The set of fields around delivery control that are interesting for a buyer to see but are non-negotiable. These are set by the publisher.

description: Option<String>

Description for the deal terms.

display_name: Option<String>

The name of the deal.

external_deal_id: Option<String>

Output only. The external deal ID assigned to this deal once the deal is finalized. This is the deal ID that shows up in serving/reporting etc.

is_setup_complete: Option<bool>

Output only. True, if the buyside inventory setup is complete for this deal.

programmatic_creative_source: Option<String>

Output only. Specifies the creative source for programmatic deals. PUBLISHER means creative is provided by seller and ADVERTISER means creative is provided by buyer.

proposal_id: Option<String>

Output only. ID of the proposal that this deal is part of.

seller_contacts: Option<Vec<ContactInformation>>

Output only. Seller contact information for the deal.

syndication_product: Option<String>

The syndication product associated with the deal. Note: This field may be set only when creating the resource. Modifying this field while updating the resource will result in an error.

targeting: Option<MarketplaceTargeting>

Output only. Specifies the subset of inventory targeted by the deal.

targeting_criterion: Option<Vec<TargetingCriteria>>

The shared targeting visible to buyers and sellers. Each shared targeting entity is AND'd together.

update_time: Option<String>

Output only. The time when the deal was last updated.

web_property_code: Option<String>

The web property code for the seller copied over from the product.