Struct google_adexchangebuyer2_v2_beta1::BidderMethods

pub struct BidderMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on bidder resources. It is not used directly, but through the AdExchangeBuyerII hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangebuyer2_v2_beta1 as adexchangebuyer2_v2_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adexchangebuyer2_v2_beta1::AdExchangeBuyerII;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeBuyerII::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `accounts_filter_sets_bid_metrics_list(...)`, `accounts_filter_sets_bid_response_errors_list(...)`, `accounts_filter_sets_bid_responses_without_bids_list(...)`, `accounts_filter_sets_create(...)`, `accounts_filter_sets_delete(...)`, `accounts_filter_sets_filtered_bid_requests_list(...)`, `accounts_filter_sets_filtered_bids_creatives_list(...)`, `accounts_filter_sets_filtered_bids_details_list(...)`, `accounts_filter_sets_filtered_bids_list(...)`, `accounts_filter_sets_get(...)`, `accounts_filter_sets_impression_metrics_list(...)`, `accounts_filter_sets_list(...)`, `accounts_filter_sets_losing_bids_list(...)`, `accounts_filter_sets_non_billable_winning_bids_list(...)`, `filter_sets_bid_metrics_list(...)`, `filter_sets_bid_response_errors_list(...)`, `filter_sets_bid_responses_without_bids_list(...)`, `filter_sets_create(...)`, `filter_sets_delete(...)`, `filter_sets_filtered_bid_requests_list(...)`, `filter_sets_filtered_bids_creatives_list(...)`, `filter_sets_filtered_bids_details_list(...)`, `filter_sets_filtered_bids_list(...)`, `filter_sets_get(...)`, `filter_sets_impression_metrics_list(...)`, `filter_sets_list(...)`, `filter_sets_losing_bids_list(...)` and `filter_sets_non_billable_winning_bids_list(...)`
// to build up your call.
let rb = hub.bidders();

Methods

impl<'a, C, A> BidderMethods<'a, C, A>

pub fn accounts_filter_sets_get(
&self,
name: &str
) -> BidderAccountFilterSetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the requested filter set for the account with the given account ID.

Arguments

name - Full name of the resource being requested. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_filtered_bids_list(
&self,
filter_set_name: &str
) -> BidderFilterSetFilteredBidListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons for which bids were filtered, with the number of bids filtered for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_list(
&self,
owner_name: &str
) -> BidderFilterSetListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all filter sets for the account with the given account ID.

Arguments

ownerName - Name of the owner (bidder or account) of the filter sets to be listed. For example: - For a bidder-level filter set for bidder 123: bidders/123 - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123 - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456

pub fn accounts_filter_sets_create(
&self,
request: FilterSet,
owner_name: &str
) -> BidderAccountFilterSetCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates the specified filter set for the account with the given account ID.

Arguments

request - No description provided.

ownerName - Name of the owner (bidder or account) of the filter set to be created. For example: - For a bidder-level filter set for bidder 123: bidders/123 - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123 - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456

pub fn accounts_filter_sets_bid_metrics_list(
&self,
filter_set_name: &str
) -> BidderAccountFilterSetBidMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all metrics that are measured in terms of number of bids.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn accounts_filter_sets_filtered_bids_creatives_list(
&self,
filter_set_name: &str,
creative_status_id: i32
) -> BidderAccountFilterSetFilteredBidCreativeListCall<'a, C, A>

Create a builder to help you perform the following task:

List all creatives associated with a specific reason for which bids were filtered, with the number of bids filtered for each creative.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

creativeStatusId - The ID of the creative status for which to retrieve a breakdown by creative. See creative-status-codes.

pub fn accounts_filter_sets_delete(
&self,
name: &str
) -> BidderAccountFilterSetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the requested filter set from the account with the given account ID.

Arguments

name - Full name of the resource to delete. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn accounts_filter_sets_filtered_bid_requests_list(
&self,
filter_set_name: &str
) -> BidderAccountFilterSetFilteredBidRequestListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons that caused a bid request not to be sent for an impression, with the number of bid requests not sent for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_filtered_bids_creatives_list(
&self,
filter_set_name: &str,
creative_status_id: i32
) -> BidderFilterSetFilteredBidCreativeListCall<'a, C, A>

Create a builder to help you perform the following task:

List all creatives associated with a specific reason for which bids were filtered, with the number of bids filtered for each creative.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

creativeStatusId - The ID of the creative status for which to retrieve a breakdown by creative. See creative-status-codes.

pub fn filter_sets_get(&self, name: &str) -> BidderFilterSetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the requested filter set for the account with the given account ID.

Arguments

name - Full name of the resource being requested. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn accounts_filter_sets_list(
&self,
owner_name: &str
) -> BidderAccountFilterSetListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all filter sets for the account with the given account ID.

Arguments

ownerName - Name of the owner (bidder or account) of the filter sets to be listed. For example: - For a bidder-level filter set for bidder 123: bidders/123 - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123 - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456

pub fn accounts_filter_sets_bid_response_errors_list(
&self,
filter_set_name: &str
) -> BidderAccountFilterSetBidResponseErrorListCall<'a, C, A>

Create a builder to help you perform the following task:

List all errors that occurred in bid responses, with the number of bid responses affected for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_delete(
&self,
name: &str
) -> BidderFilterSetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the requested filter set from the account with the given account ID.

Arguments

name - Full name of the resource to delete. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn accounts_filter_sets_impression_metrics_list(
&self,
filter_set_name: &str
) -> BidderAccountFilterSetImpressionMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all metrics that are measured in terms of number of impressions.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_losing_bids_list(
&self,
filter_set_name: &str
) -> BidderFilterSetLosingBidListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons for which bids lost in the auction, with the number of bids that lost for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn accounts_filter_sets_filtered_bids_details_list(
&self,
filter_set_name: &str,
creative_status_id: i32
) -> BidderAccountFilterSetFilteredBidDetailListCall<'a, C, A>

Create a builder to help you perform the following task:

List all details associated with a specific reason for which bids were filtered, with the number of bids filtered for each detail.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

creativeStatusId - The ID of the creative status for which to retrieve a breakdown by detail. See creative-status-codes. Details are only available for statuses 10, 14, 15, 17, 18, 19, 86, and 87.

pub fn accounts_filter_sets_filtered_bids_list(
&self,
filter_set_name: &str
) -> BidderAccountFilterSetFilteredBidListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons for which bids were filtered, with the number of bids filtered for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_filtered_bid_requests_list(
&self,
filter_set_name: &str
) -> BidderFilterSetFilteredBidRequestListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons that caused a bid request not to be sent for an impression, with the number of bid requests not sent for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn accounts_filter_sets_bid_responses_without_bids_list(
&self,
filter_set_name: &str
) -> BidderAccountFilterSetBidResponsesWithoutBidListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons for which bid responses were considered to have no applicable bids, with the number of bid responses affected for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn accounts_filter_sets_losing_bids_list(
&self,
filter_set_name: &str
) -> BidderAccountFilterSetLosingBidListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons for which bids lost in the auction, with the number of bids that lost for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_non_billable_winning_bids_list(
&self,
filter_set_name: &str
) -> BidderFilterSetNonBillableWinningBidListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons for which winning bids were not billable, with the number of bids not billed for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_impression_metrics_list(
&self,
filter_set_name: &str
) -> BidderFilterSetImpressionMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all metrics that are measured in terms of number of impressions.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_filtered_bids_details_list(
&self,
filter_set_name: &str,
creative_status_id: i32
) -> BidderFilterSetFilteredBidDetailListCall<'a, C, A>

Create a builder to help you perform the following task:

List all details associated with a specific reason for which bids were filtered, with the number of bids filtered for each detail.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

creativeStatusId - The ID of the creative status for which to retrieve a breakdown by detail. See creative-status-codes. Details are only available for statuses 10, 14, 15, 17, 18, 19, 86, and 87.

pub fn filter_sets_bid_response_errors_list(
&self,
filter_set_name: &str
) -> BidderFilterSetBidResponseErrorListCall<'a, C, A>

Create a builder to help you perform the following task:

List all errors that occurred in bid responses, with the number of bid responses affected for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn accounts_filter_sets_non_billable_winning_bids_list(
&self,
filter_set_name: &str
) -> BidderAccountFilterSetNonBillableWinningBidListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons for which winning bids were not billable, with the number of bids not billed for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_bid_metrics_list(
&self,
filter_set_name: &str
) -> BidderFilterSetBidMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all metrics that are measured in terms of number of bids.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_bid_responses_without_bids_list(
&self,
filter_set_name: &str
) -> BidderFilterSetBidResponsesWithoutBidListCall<'a, C, A>

Create a builder to help you perform the following task:

List all reasons for which bid responses were considered to have no applicable bids, with the number of bid responses affected for each reason.

Arguments

filterSetName - Name of the filter set that should be applied to the requested metrics. For example: - For a bidder-level filter set for bidder 123: bidders/123/filterSets/abc - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123/filterSets/abc - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456/filterSets/abc

pub fn filter_sets_create(
&self,
request: FilterSet,
owner_name: &str
) -> BidderFilterSetCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates the specified filter set for the account with the given account ID.

Arguments

request - No description provided.

ownerName - Name of the owner (bidder or account) of the filter set to be created. For example: - For a bidder-level filter set for bidder 123: bidders/123 - For an account-level filter set for the buyer account representing bidder 123: bidders/123/accounts/123 - For an account-level filter set for the child seat buyer account 456 whose bidder is 123: bidders/123/accounts/456